Feat : 약간의 수정

+++ app/src/main/java/com/kuit/conet/data/dto/request/member/RequestPostBookmark.kt
@@ -6,5 +6,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class RequestPostBookmark(
     @SerialName("teamId")
-    val teamId: Int,
+    val teamId: Long,
 )
